+++ src/server/lib/spreadsheet.js
@@ -28,7 +28,7 @@ const { removeMetadata } = require('../lib/remove-metadata')
     ...
   ]
   */
-function loadSpreadsheet (filename) {
+function loadSpreadsheet(filename) {
   const workbook = XLSX.readFile(filename)
   return workbook.SheetNames.map(sheetName => {
     const sheet = workbook.Sheets[sheetName]
@@ -43,7 +43,7 @@ function loadSpreadsheet (filename) {
   (not really JSON). So the first element in the array will be an array
   of column names
   */
-function sheetToJson (sheet, toLower = true) {
+function sheetToJson(sheet, toLower = true) {
   const jsonSheet = XLSX.utils.sheet_to_json(sheet, {
     header: 1,
     blankrows: false
@@ -68,7 +68,7 @@ function sheetToJson (sheet, toLower = true) {
   workbook match those of the reference template, and returns the contents
   of the uploaded workbook in a {<sheetName>: <two dimensional array>, ...}.
   */
-function parseSpreadsheet (workbook, templateSheets) {
+function parseSpreadsheet(workbook, templateSheets) {
   const valog = []
 
   // fix the sheet names
@@ -87,13 +87,17 @@ function parseSpreadsheet (workbook, templateSheets) {
   // remove metadata rows & normalize column names
   const parsedWorkbook = {}
   Object.entries(jsonWorkbook).forEach(([sheetName, sheet]) => {
+    // console.log(sheetName)
     const normalizedSheet = removeMetadata(sheetName, jsonWorkbook[sheetName])
     const [headerRow, ...rest] = normalizedSheet
+    // console.log(headerRow)
     const normalizedHeader = _.map(headerRow, colName =>
       typeof colName === 'string' ? colName.toLowerCase().trim() : colName
     )
     parsedWorkbook[sheetName] = [normalizedHeader, ...rest]
   })
+
+  // console.log(parsedWorkbook)
 
   _.forIn(templateSheets, (templateSheet, sheetName) => {
     const workbookSheet = parsedWorkbook[sheetName]
@@ -140,7 +144,7 @@ function parseSpreadsheet (workbook, templateSheets) {
                   not written to the database
     }
   */
-async function spreadsheetToDocuments (
+async function spreadsheetToDocuments(
   spreadsheet, // { <sheet name>:<two dimensional array>, ... }
   user_id,
   templateSheets
@@ -150,20 +154,30 @@ async function spreadsheetToDocuments (
 
   _.forIn(templateSheets, (templateSheet, type) => {
     const sheet = spreadsheet[type]
+
     // This case noted as a validation error in `parseSpreadsheet`
     // but allow for checking of additional errors.
     if (!sheet) return
 
+    // const sheet = removeMetadata(type, inputSheet)
     const sheetName = type.trim().toLowerCase()
+
     // Mark any columns not in the template to be ignored
     const cols = sheet[0].map(col => {
       return templateSheet[0].includes(col) ? col : 'ignore'
     })
+    // console.log('\n\n')
+    // console.log(type)
+    // console.log(templateSheet[0])
+    // console.log(sheet[0])
+    // console.log(cols)
+
     // Process data rows
     sheet.slice(1).forEach((row, i) => {
       if (row.length === 0) return
 
       const jsonRow = _.omit(_.zipObject(cols, row), ['ignore'])
+      // console.log(jsonRow)
 
       switch (sheetName) {
         case 'logic':
@@ -172,8 +186,40 @@ async function spreadsheetToDocuments (
           // omit these sheets
           return
 
+        // case 'subrecipient':
+        //   if (jsonRow['duns number']) {
+        //     // populate the identification number field for easier deduplication
+        //     jsonRow['identification number'] = `DUNS${jsonRow['duns number']}`
+        //   }
+        //   break
+
+        // case 'projects':
+        //   jsonRow['project identification number'] =
+        //     zeroPad(jsonRow['project identification number'])
+        //   break
+
+        // case 'contracts':
+        // case 'grants':
+        // case 'loans':
+        // case 'transfers':
+        // case 'direct':
+        // case 'cover':
+        //   // note in the projects database table this field is called "code",
+        //   // and "id" is not this, but the numeric record id.
+        //   jsonRow['project id'] = zeroPad(jsonRow['project id'])
+        //   break
+
+        // case 'certification':
+        // case 'aggregate awards < 50000':
+        // case 'aggregate payments individual':
+        //   break
+
+
         default:
           break
+        // default:
+        //   // ignore all other sheets
+        //   return
       }
 
       documents.push({
@@ -191,20 +237,20 @@ async function spreadsheetToDocuments (
   record to preserve the source row for validation reporting. We need to
   get rid of it before attempting to write the document to the db
   */
-function removeSourceRowField (documents) {
+function removeSourceRowField(documents) {
   return documents.map(document => {
     delete document.sourceRow
     return document
   })
 }
 
-function uploadFilename (filename) {
+function uploadFilename(filename) {
   return `${process.env.UPLOAD_DIRECTORY}/${filename}`
 }
 
 /* clean() trims strings and rounds amounts
-  */
-function clean (objRecord) {
+ */
+function clean(objRecord) {
   const objCleaned = {}
   Object.keys(objRecord).forEach(key => {
     let val = objRecord[key] || null
@@ -214,7 +260,7 @@ function clean (objRecord) {
     }
     switch (columnTypeMap[key]) {
       case 'amount':
-        objCleaned[key] = _.round((Number(val) || 0), 2) || null
+        objCleaned[key] = _.round(Number(val) || 0, 2) || null
         break
 
       case 'string': {
@@ -242,23 +288,24 @@ function clean (objRecord) {
 // not remove a trailing or leading double quote without a corresponding
 // one at the other end).
 // - Converts double spaces to single spaces.
-function cleanString (val) {
+function cleanString(val) {
   if (!val && val !== 0) {
     return null
   }
   val = String(val).trim()
   if (val) {
-    val = val.replace(/^"(.+)"$/, '$1')
+    val = val
+      .replace(/^"(.+)"$/, '$1')
       .replace(/ {2}/g, ' ')
       .trim()
   }
   return val
 }
 
-function zeroPad (code) {
+function zeroPad(code) {
   code = String(code)
   if (code.length < 3) {
-    code = (`000${code}`).substr(-3)
+    code = `000${code}`.substr(-3)
   }
   return code
 }
